+++ example/convex/_generated/api.d.ts
@@ -41,4 +41,46 @@ export declare const internal: FilterApi<
   FunctionReference<any, "internal">
 >;
 
+export declare const components: {
+  launchdarkly: {
+    events: {
+      storeEvents: FunctionReference<
+        "mutation",
+        "internal",
+        { eventsUri?: string; payloads: Array<string>; sdkKey: string },
+        null
+      >;
+    };
+    store: {
+      get: FunctionReference<
+        "query",
+        "internal",
+        { key: string; kind: "flags" | "segments" },
+        string | null
+      >;
+      getAll: FunctionReference<
+        "query",
+        "internal",
+        { kind: "flags" | "segments" },
+        Array<string>
+      >;
+      initialized: FunctionReference<"query", "internal", {}, boolean>;
+      write: FunctionReference<
+        "mutation",
+        "internal",
+        { payload: string },
+        null
+      >;
+    };
+    tokens: {
+      validate: FunctionReference<
+        "query",
+        "internal",
+        { token?: string },
+        { error?: string; success: boolean }
+      >;
+    };
+  };
+};
+
 /* prettier-ignore-end */
