+++ example/convex/tsconfig.json
@@ -18,7 +18,12 @@
     "forceConsistentCasingInFileNames": true,
     "module": "ESNext",
     "isolatedModules": true,
-    "noEmit": true
+    "noEmit": true,
+    /* This should only be used in this example. Real apps should not attempt
+     * to compile TypeScript because differences between tsconfig.json files can
+     * cause the code to be compiled differently.
+     */
+     "customConditions": ["@convex-dev/component-source"]
   },
   "include": ["./**/*", "../launchdarkly/schema.ts"],
   "exclude": ["./_generated"]
