+++ example/package.json
@@ -5,14 +5,14 @@
   "type": "module",
   "scripts": {
     "dev": "vite",
-    "cvx": "convex dev --live-component-sources",
+    "cvx": "convex dev --live-component-sources --typecheck-components",
     "build": "tsc -b && vite build",
     "lint": "eslint . --ext ts,tsx --report-unused-disable-directives --max-warnings 0",
     "preview": "vite preview"
   },
   "dependencies": {
-    "convex": "1.16.0-alpha.2",
-    "launchdarkly-component": "file:../",
+    "convex": "^1.16.5",
+    "@convex-dev/launchdarkly": "file:../",
     "react": "^18.3.1",
     "react-dom": "^18.3.1"
   },
