+++ package.json
@@ -1,5 +1,7 @@
 {
   "name": "@convex-dev/launchdarkly",
+  "description": "A Convex component that syncs LaunchDarkly feature flags.",
+  "homepage": "https://github.com/get-convex/launchdarkly#readme",
   "version": "0.0.1",
   "type": "module",
   "repository": "github:get-convex/launchdarkly",
@@ -27,9 +29,11 @@
     "dist",
     "src"
   ],
+  "peerDependencies": {
+    "convex": "~1.16.5 || ~1.17.0"
+  },
   "dependencies": {
     "@launchdarkly/js-server-sdk-common": "2.6.1",
-    "convex": "^1.16.1",
     "crypto-js": "^4.2.0"
   },
   "devDependencies": {
@@ -46,15 +50,17 @@
     "./package.json": "./package.json",
     ".": {
       "import": {
+        "@convex-dev/component-source": "./src/component/index.ts",
         "types": "./dist/esm/index.d.ts",
         "default": "./dist/esm/index.js"
       },
       "require": {
+        "@convex-dev/component-source": "./src/component/index.ts",
         "types": "./dist/commonjs/index.d.ts",
         "default": "./dist/commonjs/index.js"
       }
     },
-    "./convex.config.js": {
+    "./convex.config": {
       "import": {
         "@convex-dev/component-source": "./src/component/convex.config.ts",
         "types": "./dist/esm/component/convex.config.d.ts",
