+++ src/index.ts
@@ -1,4 +1,2 @@
 export { registerRoutes } from "./component/registerRoutes";
 export { LDClient } from "./sdk/LDClient";
-import launchdarkly from "./component/convex.config";
-export default launchdarkly;
